+++ crates/plugin-api/CHANGELOG.md
@@ -10,7 +10,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed 
 
-- `initialize_runtime` accepts a `javy_plugin_api::Config` instead of a `javy_plugin_api::javy::Config`
+- `initialize_runtime` accepts a `javy_plugin_api::Config` instead of a
+  `javy_plugin_api::javy::Config`
+
+### Added
+
+- Can now enable the event loop using the `javy_plugin_api::Config`
 
 ## [1.0.0] - 2024-11-12